Type checking cop with simple literal methods
A basis for the fundamental operations of type checking StrongRuby, and
possible subclasses that refine and specialize the process.

AST nodes are extended with a typing context, left entirely to the
cop(s) to manage. One or several passes through the AST populate the
nodes with typing information, compose this information and perform
various type checks. Operations may be required before or after visiting
a node's children, and so the inclusion of Traversal is private to the
cop, and on_ methods are re-used there, unbeknownst to the commissioner,
and pre- and post-processing tailored in a transparent manner.

Type checking will be implemented constructively, with the first check
involving propagation of values from constants to simple methods and
checking against declared return types.

No effort is currently planned to specialize away from StrongRuby, with
regular Ruby being on the extreme where no type information avails.
Target versions must be adjusted accordingly, and
lib/rubocop/rspec/cop_helper.rb point to StrongRuby instead of regular
Ruby.