Typing of method arguments
Method arguments initialize the local context with their typing
information, implicit or explicit. Processing proceeds sequentially on
the args node, reproducing there a restricted version of the recursion
on child nodes. It would be more laborious to do this on annot, which is
used for both argument types and return types, and each arg is not aware
on its own of whether it is annotated or not.