Typing context in conditional statements
The branching and merging of the typing context in the context of
conditionals deviates from the (usually convenient) sequential model. To
address it, we replace the regular traversal, delegated to the
superclass, with an extension that adds context bookkeeping, creating a
fresh copy of the context before each branch and merging the resulting
contexts together, after the sequential processing of each.

As per established convention, typing information is merged when both
halves are found to exist and be in agreement: both the absence of a
variable in one half or disagreement in types results in exclusion from
the final context. There will always be two "branches" to merge: a
single-branch conditional will be merged with the base, unmodified
context, whereas a two-branch conditional will merge those, both derived
from the base.