Typing of .new and object inheritance
Creation of new objects through the .new method is represented by
sending the message :new to the constant representing the name of the
class. The message is special in the sense that typing is known without
annotations, unlike other messages for which no such implicit knowledge
avails.

Currently, only simple class names propagate the "typing" of const
nodes, and this is propagated to the send node when possible.