Return typing of local method calls
Initial treatment of the `send` primitive for non-constructors. A first
approximation, based on a conceptually flat namespace and everything
happening inside the same "class" or unit, stores the root node of the
AST and searches all method signatures when a message is sent by the
program. Each `def` node represents a method definition and has
associated a return type: whether explicitly given or, implicitly,
Object (we have some minimal typing information, and this is more than
no type information).

Note the AST is not available during `initialize` and the class variable
that holds it can only be set from `investigate`.

Clearly, this model is as grossly simplified as it is inefficient. To
begin with, the rich structure of modules, classes, receivers of static
and instance methods, etc., needs to be gradually built into the
process. What's more, method signatures and the aforementioned structure
are prime targets for caching within a single run of a cop, but also
cumulatively across different invocations to the type checking cop
within a single checking session (and beyond).

As these processes become more complex and come to represent preliminary
passes of the checking process, walker classes can be furnished to
encapsulate the necessary functionality.

This commit considers return types exclusively.