Arity of local method calls with simple arguments
Under the same set of simplified assumptions that currently cover local
method calls, a first check verifies that the arity of the call is,
indeed, that of the method definition. The code is slightly more complex
than it need be to cover arity, strictly speaking, but retrieves the
necessary typing information for the natural consequence of a successful
arity check: typechecking the concrete arguments of each call.

Some restructuring in the code keeps methods small and modular, although
outstanding refactoring possibilities are noted, as usual.

This commit represents a brief departure from the principle that RuboCop
must run cleanly on RuboCop before committing any changes, if only
because the complexity of argument lists goes beyond the increment
presented herein. Further refinements to account for optional arguments
and other variations will presently resolve this situation.