Typing of simple method arguments
Given two arrays of nodes representing the parameters of a
message/method call and the types of the formal arguments to the
corresponding method, both of the same length, if the parameter nodes
have already been evaluated, check that their types are compatibles with
those of the signature, and generate an error message if not.

Given the current scope of flat namespaces and single sources, it will
be common for the subtyping relation to be partially defined. In
practice, we need to recover from NameErrors whenever we fail to
retrieve Class objects from their symbolic names. This will not suffice
to retrieve the relations described in ASTs and unloaded source, but
keeps RuboCop's runs clean and unaffected by the embrionary state of the
type checking cop(s).

This does not mean that the runs are free of offenses; in fact, a great
many are found in code that is not yet fully understood. These are to be
expected, and a sign that everything starts to fall in place. Following
commits will treat many of these "errors" by refining the capabilities
of the type checker.