Typing of optional arguments
When annotated, optional arguments adopt the type of the annotation over
that of their default value, although the latter must be compatible with
(i.e., a subclass of) the former. Otherwise, a typing error is
generated, and subsequent errors on call sites relying on this defective
default value can be expected. The following extensions will complete
the semantics of optional arguments that is introduced here.

In addition, and given sufficient type checking information, this check
generalizes and supersedes the proof-of-concept typechecking cop,
DefaultLiteral.