Typing of optional defaults and arity of calls
A short and clear path to implementation computes an array of optional
argument indexes in parallel with the types of each argument, computes
and compares against a range.

Type annotations on optional parameters must be consistent with default
values (and here the inference on general expressions will need to pass
more stringent tests).

No effort is made to restrict the types of unannotated optional
parameters based on their default values, although it is a possibility
that can be contemplated.