Typing of local method calls with optargs
Once it has been ascertained that the number of arguments passed to a
method with optional parameters fall in the accepted range, it is
necessary to match the actual arguments with the formal parameters. To
do this, we excise those types from the list of argument types retrieved
for the method that correspond to those optional parameters that will
take their default values from the definition.

These default parameters will not be typechecked at this point: doing so
would require at this point some out-of-order processing of ASTs that
would be inconvenient; a more general concern would be that a typing
error will be generated when parsing the corresponding method
definition, if the default value does not match the type annotation, if
any. By avoiding this check here, we will ensure that any such errors
will become effective exactly once, as opposed to in the declaration and
at all call sites that rely on defective defaults.

Finally, some code is moved around and refactored for greater clarity.