Refactor local method signatures
Two new classes begin to abstract the representation and management of
types and signatures in StrongRuby. No new functionality is added. The
focus is on encapsulating the relevant typing information that describes
a method definition, as well as preprocessing the source and collecting
all this information, currently limited to the local file, in a
central repository, accessible everywhere in the analysis and guaranteed
to be complete before the actual typechecking begins.

This is both necessary for and prepared to support multi-file and
complex organizations of code, which will be developed in subsequent
commits.