Visibility of source-local methods by class
Support for contexts and namespaces begins with the basic relation of
methods defined in certain classes, and available to objects of those
classes, for now limited to a single processed source (for the simple
reason that signatures are currently read for the current source only,
for lack of environment information about the set of source files in
scope). Functionally, this corresponds to the following pair of matching
extensions:

 - Signature traversals must now keep track of (for now, only) the
   current class scope, so that method signatures will always be
   assigned to the class where they are defined.

   (An interesting case is that of methods defined outside a class, as
   seen in some tests, where currently the initial, null class scope is
   used.)

 - When sending a message, previous checks on argument and return types
   are performed only when the signature can determine that a method
   named like the message has been defined for the class to which the
   receiver object belongs. The class of an object is determined using
   local type inference in the usual way. If a matching method is not
   found, an offense is added.

On the testing side of things, some precisions are necessary.

 - Thorough verification of method definition requires, in general,
   access to the Ruby runtime, which will cause ample numbers of
   warnings for code relying on the standard library, due to the
   dichotomy that exists between what is visible by the runtime and what
   is visible by the signature.

 - The former point only affects tests for the type checker in that
   conditions involving comparisons are simplified so as not to rely on
   any methods in the default environment.

 - In addition, some new error messages are added to existing tests that
   rely on the propagated (and undefined) values of undefined calls (or
   variables, both categories being syntactically ambiguous). The end
   result, i.e., that the resulting value of propagating an undefined
   value is itself undefined, is joined by the "root cause" of it,
   namely the statement that originates the undefined value.

It is important to note that until both runtime and signature worlds are
reconciled, the number of offenses that will be found in code for which
part of the environment is not available grows very considerably with
the present additions.

Exacerbated by these developments, but likely beneficial whatever the
cases at the current stage, it is recommended to remove the type checker
from the set of cops enabled by default, so that tests that rely on runs
of this configuration will continue to pass, as they would fail if
additional typing errors were found by the evolving checker.