Simple class scope and local method calls
The type checker is made aware of classes as it walks the AST, and is
therefore capable of deciding where to look for a method when an
explicit receiver is given, i.e., it knows in which class it is.

Preliminary conventions are made for nested classes, although these are
not considered in this commit. A canonical encoding of namespaces, and
the abstraction thereof, has to be considered as well.